requirements.txt
# =============================================================================
# ImmunePlan Pipeline Requirements
# =============================================================================
# Note: PyTorch installed separately via setup.sh with CUDA support
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
# =============================================================================

# --- Core ML ---
transformers>=4.40.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
peft>=0.10.0
datasets>=2.18.0

# --- Training ---
sentencepiece>=0.2.0
protobuf>=4.25.0
tensorboard>=2.16.0

# --- Data Collection ---
arxiv>=2.1.0
biopython>=1.83
requests>=2.31.0
beautifulsoup4>=4.12.0
pymupdf>=1.24.0

# --- RAG & Embeddings ---
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.26.0

# --- API ---
Flask>=3.0.0
flask-cors>=4.0.0

# --- Data Processing ---
pandas>=2.2.0
scikit-learn>=1.4.0

# --- Environment ---
python-dotenv>=1.0.0

# --- GPU Monitoring ---
gputil>=1.4.0
psutil>=5.9.0

# --- Pipeline (optional) ---
# zenml>=0.50.0
